add enum description for 3PInverterData - and fix datatype bugs

+++ FroniusSolarClient/Entities/SolarAPI/V1/InverterRealtimeData/CommonInverterData.cs
@@ -13,7 +13,7 @@ namespace FroniusSolarClient.Entities.SolarAPI.V1.InverterRealtimeData
         /// AC power 
         /// </summary>
         [JsonProperty("PAC")]
-        public UnitValue<int> AcPower { get; set; }
+        public UnitValue<decimal> AcPower { get; set; }
 
         /// <summary>
         /// AC current
